### 2017-12-06 #485 @whedon
[PRE REVIEW]: pyodesys: Straightforward numerical integration of ODE systems from Python
**Submitting author:** @bjodah (<a href="http://orcid.org/0000-0003-0596-0222">Björn Dahlgren</a>)
**Repository:** <a href="https://github.com/bjodah/pyodesys" target ="_blank">https://github.com/bjodah/pyodesys</a>
**Version:** v0.11.5
