### 2017-12-06 #485 @arfon
👋 @bjodah - looks like there are some issues with pandoc and your bibtex file, perhaps with the LaTeX syntax on [this line](https://github.com/bjodah/pyodesys/blob/master/joss-paper/paper.bib#L46)? Perhaps these can just be replaced with `+` symbols rather than LaTeX commands?
