### 2017-12-06 4a71c0c
Update paper.bib
