+++ joss-paper/paper.bib
@@ -43,5 +43,5 @@
   year  = {2011},
   publisher = {{AIP}},
   author = {Karsten Ahnert and Mario Mulansky and Theodore E. Simos and George Psihoyios and Ch. Tsitouras and Zacharias Anastassi},
-  title = {Odeint {\textendash} Solving Ordinary Differential Equations in C$\mathplus$$\mathplus$}
+  title = {Odeint - Solving Ordinary Differential Equations in C++}
 }
