### 2017-12-06 #485 @arghdos
@bjodah Looks interesting, I'd be happy to review
