### 2017-12-07 8d877cc
Better wording in JOSS-manuscript
