### 2017-12-15 bjodah/pyodesys#67 @arghdos
JOSS: Issues installing
@bjodah I'm having some issues installing `pyodesys` for the JOSS review:

First, let me note that installing the main package via conda worked without a hitch, pip not so much (but that's a more difficult nut to crack).  However:

1.  There needs to be a list of _all_ dependencies that are needed to run tests / examples / etc.  A lot of this can be inferred from either the setup.py or the other optional packages, but I tried that route and I'm still running into errors on tests / examples (more on this later).  Preferably this list would be the docs, and there would be a link in the install section along the lines of "Optional Dependencies...".  Even if this links to the various subpackages (e.g. [pygslodeiv2](https://github.com/bjodah/pygslodeiv2)) which then tell you how to install themselves, that would be an improvement.

2.  This list needs to include any non-python packages required such that a pip install will be possible.
 Conda is obviously an easier choice, but if you're going to offer pip installation I need to the base packages required.  Again using pygslodeiv2 as an example, it's not immediately obvious where GSL should be installed to (other than "default" locations?).  Or for [odeint](http://headmyshoulder.github.io/odeint-v2/) do I need Boost? ode-int itself?

3.  Even with (I think) all the sub-dependencies installed, most of the tests skip and the examples give erros.  My conda environment is [here](https://gist.github.com/arghdos/fd9f012cb10278e0de282cf55a5d5544)

e.g.: my test output:

> pyodesys/pyodesys/native/tests/test_cvode.py sssssssssssssssssssssssssssss                                                                                       [ 11%]
> pyodesys/pyodesys/native/tests/test_gsl.py sssssFsssssssss                                                                                                       [ 16%]
> pyodesys/pyodesys/native/tests/test_odeint.py ssssFsssssss                                                                                                       [ 21%]
> pyodesys/pyodesys/tests/test_core.py s.ssssssss....s.....s..s..s..ssss.                                                                                          [ 34%]
> pyodesys/pyodesys/tests/test_plotting.py ss                                                                                                                      [ 35%]
> pyodesys/pyodesys/tests/test_results.py ..s                                                                                                                      [ 36%]
> pyodesys/pyodesys/tests/test_robertson.py ssssssssss                                                                                                             [ 40%]
> pyodesys/pyodesys/tests/test_symbolic.py ....ssssssssss.................ssssssssssssssssssssssssssssssssssssssssss...sssssssssss.sss.s.s....ss..ssssssssssssssss [ 86%]
> sssssssssssssssssssssss.ssssssssss                                                                                                                               [ 99%]
> pyodesys/pyodesys/tests/test_util.py ..

Or running `chemical_kinetics.ipynb`:

> ~/miniconda3/envs/pyodesys/lib/python3.6/site-packages/pycvodes/__init__.py in <module>()
>       7 import numpy as np
>       8
> ----> 9 from ._cvodes import adaptive, predefined, requires_jac, steppers, fpes
>      10 from ._util import _check_callable, _check_indexing
>      11 from ._release import __version__
>
> ImportError: libsundials_cvodes.so.2: cannot open shared object file: No such file or directory

edit: link back to joss review:
https://github.com/openjournals/joss-reviews/issues/490#issuecomment-352495520
