### 2017-12-15 bjodah/pyodesys#67 @bjodah
@arghdos, thank you for your thorough feedback on the installation issues. Right now for installation using pip I've taken an "all or nothing" approach (where `pip install pyodesys` installs the bare minimum and `pip install pyodesys[all]` installs everything). The original reason for this was that the 3rd party libraries were not always easily accessible on windows (or even OS X) until recently (now they are all available for all platforms via conda).

Arguably I could extend what the "bare minimum" is (e.g. have it include also `sympy`, `scipy`, `matplotlib` & `pytest`) for `pyodesys`. Then a basic "pip-install" of the package will be more useful. That would also lead to fewer skipped tests due to missing imports (maybe I should also instruct the user to pass the `-rs` flag to pytest?). Do you think that would be an improvement?

- `1/2)` I'll add an "Optional dependencies" subsection to the installation section. I'll also add a note about what 3rd party libraries those include (sundials 2.7.0/gsl >=1.16/boost >=1.65.0) for pycvodes/pygslodeiv2/pyodeint respectively), in addition I can add the required package names in e.g. debian/ubuntu.

- `3)` I could add a link to a binder (which would use a Dockerimage)? That would then serve both as an example of setup as well as an easy way for prospect users to explore the package.
