### 2017-12-15 c96a898
Elaborate on dependencies in README
