+++ setup.py
@@ -82,11 +82,10 @@ if not len(long_descr) > 100:
 _author, _author_email = open(_path_under_setup('AUTHORS'), 'rt').readline().split('<')
 
 extras_req = {
-    'integrators': ['scipy>=0.19.1', 'pyodeint>=0.9.0', 'pycvodes>=0.9.0', 'pygslodeiv2>=0.8.0'],
-    'symbolic': ['sym', 'sympy>=1.1.1'],
+    'integrators': ['pyodeint>=0.9.0', 'pycvodes>=0.9.0', 'pygslodeiv2>=0.8.0'],
     'native': ['pycompilation>=0.4.3', 'pycodeexport>=0.1.1', 'appdirs'],
     'docs': ['Sphinx', 'sphinx_rtd_theme', 'numpydoc'],
-    'testing': ['pytest', 'pytest-cov', 'pytest-flakes', 'pytest-pep8']
+    'testing': ['pytest-cov', 'pytest-flakes', 'pytest-pep8']
 }
 extras_req['all'] = list(chain(extras_req.values()))
 
@@ -102,7 +101,7 @@ setup_kwargs = dict(
     license=license,
     packages=[pkg_name] + submodules + tests,
     include_package_data=True,
-    install_requires=['numpy'],
+    install_requires=['numpy>=1.8.0', 'pytest>=2.9.2', 'scipy>=0.19.1', 'sym>=0.3.0', 'sympy>=1.1.1', 'matplotlib>=1.5.3'],
     extras_require=extras_req
 )
 
