### 2017-12-18 bjodah/pyodesys#67 @arghdos
@bjodah Ok, that helped a lot (the sundials version in particular was a hangup for me, as by default conda installs 3.x.x).  Additionally, the -rs option is also helpful as it lists some other packages that might be missing in the setup dependencies (or possibly sub-dependencies?) --`symcxx`, `quantities`, `symengine` (and `matplotlib`, but it looks like you fixed that already).

The docker image is a great idea as well! Particularly for complicated mixed python / c packages I think it's critical to make the installation as well explained as possible, it's very easy (as we're showing here!) for small details to result in broken installs.  Also an entry point into exploring the package / examples would be excellent.

I still end up with [failed tests however](https://gist.github.com/arghdos/657ff285df54827d42ea658908733124)).  It appears to be only for tests that require `pycompilation` or `pycodeexport`.  `g++ --version` on this system reports:

>g++ (Ubuntu 5.4.0-6ubuntu1~16.04.4) 5.4.0 20160609

which as far as I can tell should support c++14 as required for these examples, any thoughts here?
