### 2017-12-18 #490 @arghdos
@arfon I've been working with @bjodah to clean up the installation instructions in the documentation (I ran into some issues while installing / testing).  Once we have that finalized I expect I'll be finished shortly thereafter
