### 2017-12-19 78deb12
Use conda on binder, use Dockerfile for local script
