### 2017-12-19 61b13b6
Add notebook as dependency, add postBuild for mybinder
