### 2017-12-19 e272930
Tweak docker script, update README
