### 2017-12-19 38e59a6
Update environment.yml & index.ipynb
