### 2017-12-19 01e277f
Fix repo2docker setup. Run tests using docker
