+++ README.rst
@@ -115,10 +115,16 @@ Using Docker
 If you have `Docker <https://www.docker.com>` installed, you may use it to host a jupyter
 notebook server::
 
-  $ ./scripts/host-jupyter-using-docker.sh
+  $ ./scripts/host-jupyter-using-docker.sh . 8888
 
 the first time you run the command some dependencies will be downloaded. When the installation
-is complete there will be a link visible which you can open in your browser.
+is complete there will be a link visible which you can open in your browser. You can also run
+the test suite using the same docker-image::
+
+  $ ./scripts/host-jupyter-using-docker.sh . 0
+
+there will be one skipped test (due to symengine missing in this pip installed environment) and
+quite a few instances of RintimeWarning.
 
 Examples
 --------
