### 2017-12-19 bjodah/pyodesys#67 @bjodah
@arghdos thanks.

g++5.4 is new enough (that's the version I use myself). Looking at your latest failed test log I see:
```
...
nvector/nvector_serial.h: No such file or directory
```
this suggests that sundials' headers are not installed in any of the default include directories. One way to work around that is setting e.g. `CPLUS_INCLUDE_PATH=/path/to/sundials-2.7/include`

Regarding sundials-2.7.0 vs. 3.x: yes, right now `pycvodes` only works with 2.7.0, they broke both API and ABI moving from 2.7.0 -> 3.x. It is "just" a matter of adding a bunch of `#ifdef` blocks to pycvodes, I will do so as soon as I find some time.

I have now added both an `environment.yml` file for use with binder (link in README) and a script for hosting a docker-based jupyter notebook with pip-installed dependencies (with instructions for how to host in README). That Dockerfile then also serves as an example of how to setup a system with the required 3rd party libraries (the Dockerfile is based on the current debian stable).

Regarding `symcxx`, `quantities` and `python-symengine`:
I added the first two to the pip-installed environment in the Dockerfile. I'd rather not add symengine to that list since their official way of installing is through conda. So any tests depending on symengine (currently there is 1 such test) will be skipped in the Docker environment (it is still being run on the CI server).
