### 2017-12-19 bjodah/pyodesys#67 @arghdos
...
nvector/nvector_serial.h: No such file or directory
this suggests that sundials' headers are not installed in any of the default include directories.

>One way to work around that is setting e.g. CPLUS_INCLUDE_PATH=/path/to/sundials-2.7/include

Aaahhh, I had missed that in the noise of the log. I had gotten pycvode to install correctly previously by passing the [miniconda directories to pip](https://stackoverflow.com/questions/18783390/python-pip-specify-a-library-directory-and-an-include-directory) but it makes sense that it needs them at compile time as well.  I'll give that a shot
