### 2017-12-19 bjodah/pyodesys#67 @arghdos
@bjodah that did the trick (along with a LIBRARY_PATH to point the library search in the right direction).  I'll give the docker a test as well, but I think this issue can be closed
