### 2017-12-21 bjodah/pyodesys#69 @arghdos
JOSS: Missing Contributing Guidelines
@bjodah the only remaining issue for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/490) is a missing contribution guidlines:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

See here for [an example](https://gist.github.com/PurpleBooth/b24679402957c63ec426)
