### 2017-12-21 #490 @arghdos
@bjodah looks great! The only remaining issue is a missing contribution guidelines -- in particular, the sympy integration allowing change of variables / reformulation, and automatic analytical Jacobian determination is _very_ cool, great work
