### 2017-12-28 bjodah/pyodesys#69 @bjodah
Thanks. Yes, I should probably have written a bit more than a single sentence in the readme. I'm adressing this in gh-70.
