### 2018-01-02 bjodah/pyodesys#69 @arghdos
@bjodah looks good!  Sorry for the delay, was out of town.
