### 2018-01-02 bjodah/pyodesys#69 @bjodah
No worries, hope you enjoyed the holidays. I merged gh-70.
