### 2018-01-03 c8824f2
Prepare for v0.11.6
