+++ CHANGES.rst
@@ -1,3 +1,8 @@
+v0.11.6
+=======
+- Changed install_requirements in setup.py to be less minimalistic.
+- Improved install instructions.
+
 v0.11.5
 =======
 - Introduced ODESys.numpy to allow user to used a patched numpy module. (e.g. for quantitites)
