+++ setup.cfg
@@ -1,5 +1,5 @@
 [tool:pytest]
-norecursedirs = .git .cache scripts build dist conda-recipe benchmarks __pycache__ doc deploy
+norecursedirs = .git .cache scripts build dist conda-recipe benchmarks __pycache__ doc deploy .jupyter
 pep8maxlinelength=119
 pep8ignore =
     doc/conf.py ALL
