### 2018-01-03 16cac50
Release docs for v0.11.6
