+++ v0.11.6/_sources/index.rst.txt
@@ -1,0 +1,273 @@
+.. pyodesys documentation master file, created by
+   sphinx-quickstart on Wed Jan  3 22:27:30 2018.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to pyodesys's documentation!
+====================================
+
+.. toctree::
+   :maxdepth: 4
+   :caption: Contents:
+
+   pyodesys
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
+
+Overview
+========
+
+.. image:: http://hera.physchem.kth.se:9090/api/badges/bjodah/pyodesys/status.svg
+   :target: http://hera.physchem.kth.se:9090/bjodah/pyodesys
+   :alt: Build status on Drone
+.. image:: https://circleci.com/gh/bjodah/pyodesys.svg?style=svg
+   :target: https://circleci.com/gh/bjodah/pyodesys
+   :alt: Build status on CircleCI
+.. image:: https://img.shields.io/pypi/v/pyodesys.svg
+   :target: https://pypi.python.org/pypi/pyodesys
+   :alt: PyPI version
+.. image:: https://img.shields.io/badge/python-2.7,3.4,3.5-blue.svg
+   :target: https://www.python.org/
+   :alt: Python version
+.. image:: https://img.shields.io/pypi/l/pyodesys.svg
+   :target: https://github.com/bjodah/pyodesys/blob/master/LICENSE
+   :alt: License
+.. image:: http://img.shields.io/badge/benchmarked%20by-asv-green.svg?style=flat
+   :target: http://hera.physchem.kth.se/~pyodesys/benchmarks
+   :alt: airspeedvelocity
+.. image:: http://hera.physchem.kth.se/~pyodesys/branches/master/htmlcov/coverage.svg
+   :target: http://hera.physchem.kth.se/~pyodesys/branches/master/htmlcov
+   :alt: coverage
+.. image:: https://zenodo.org/badge/43131469.svg
+   :target: https://zenodo.org/badge/latestdoi/43131469
+   :alt: Zenodo DOI
+
+`pyodesys <https://github.com/bjodah/pyodesys>`_ provides a straightforward way
+of numerically integrating systems of ordinary differential equations (initial value problems).
+It unifies the interface of several libraries for performing the numerical integration as well as
+several libraries for symbolic representation. It also provides a convenience class for 
+representing and integrating ODE systems defined by symbolic expressions, e.g. `SymPy <http://www.sympy.org>`_
+expressions. This allows the user to write concise code and rely on ``pyodesys`` to handle the subtle differences
+between libraries.
+
+The numerical integration is performed using either:
+
+- `scipy.integrate.ode <http://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.ode.html>`_
+- `pygslodeiv2 <https://github.com/bjodah/pygslodeiv2>`_
+- `pyodeint <https://github.com/bjodah/pyodeint>`_
+- `pycvodes <https://github.com/bjodah/pycvodes>`_
+
+Note that implicit steppers require a user supplied callback for calculating the Jacobian.
+``pyodesys.SymbolicSys`` derives the Jacobian automatically.
+
+The symbolic representation is usually in the form of `SymPy <https://www.sympy.org/>`_
+expressions, but the user may choose another symbolic back-end (see `sym <https://github.com/bjodah/sym>`_).
+
+When performance is of utmost importance, e.g. in model fitting where results are needed
+for a large set of initial conditions and parameters, the user may transparently
[…]
