### 2018-01-03 92575dd
Update version for paper
