### 2018-01-03 #490 @bjodah
@arfon Thanks,

I've tagged a new patch version (v0.11.6) including the changes suggested by arghdos.
The DOI for the archive of this version is: [10.5281/zenodo.1134944](http://dx.doi.org/10.5281/zenodo.1134944)
