### 2018-01-03 #490 @bjodah
...and I just updated the paper meta-data accordingly:
https://github.com/bjodah/pyodesys/commit/92575ddf77ad795247a0f890319be0e94beacc45
