### 2018-01-04 #490 @arfon
@arghdos - many thanks for your review here ✨

@bjodah - your paper is now accepted into JOSS and your paper DOI is https://doi.org/10.21105/joss.00490 ⚡️ 🚀 💥
